pyproject.toml
[project]
name = "pytest-drf"
version = "0.1.0"
description = "Implementation of unit tests with pytest lib for Django REST Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-redis>=5.4.0",
    "djangorestframework>=3.15.2",
    "faker>=37.0.0",
    "pymemcache>=4.0.0",
    "pytest-django>=4.10.0",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = [
    "test_*.py",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
]